LayoutTests/imported/w3c/web-platform-tests/webrtc-encoded-transform/script-transform-generateKeyFrame-simulcast.https-expected.txt



PASS generateKeyFrame works with simulcast rids
PASS generateKeyFrame for rid that was negotiated away fails
FAIL generateKeyFrame with rid after simulcast->unicast negotiation fails assert_equals: expected "failure" but got "success"
